CDAP-15848 Move packages
