+++ src/main/java/io/cdap/plugin/pdf/PDFExtractor.java
@@ -14,20 +14,20 @@
  * the License.
  */
 
-package co.cask.hydrator.plugin;
+package io.cdap.plugin.pdf;
 
-import co.cask.cdap.api.annotation.Description;
-import co.cask.cdap.api.annotation.Macro;
-import co.cask.cdap.api.annotation.Name;
-import co.cask.cdap.api.annotation.Plugin;
-import co.cask.cdap.api.data.format.StructuredRecord;
-import co.cask.cdap.api.data.schema.Schema;
-import co.cask.cdap.api.plugin.PluginConfig;
-import co.cask.cdap.etl.api.Emitter;
-import co.cask.cdap.etl.api.PipelineConfigurer;
-import co.cask.cdap.etl.api.Transform;
-import co.cask.cdap.etl.api.TransformContext;
 import com.google.common.annotations.VisibleForTesting;
+import io.cdap.cdap.api.annotation.Description;
+import io.cdap.cdap.api.annotation.Macro;
+import io.cdap.cdap.api.annotation.Name;
+import io.cdap.cdap.api.annotation.Plugin;
+import io.cdap.cdap.api.data.format.StructuredRecord;
+import io.cdap.cdap.api.data.schema.Schema;
+import io.cdap.cdap.api.plugin.PluginConfig;
+import io.cdap.cdap.etl.api.Emitter;
+import io.cdap.cdap.etl.api.PipelineConfigurer;
+import io.cdap.cdap.etl.api.Transform;
+import io.cdap.cdap.etl.api.TransformContext;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDDocumentInformation;
 import org.apache.pdfbox.pdmodel.encryption.InvalidPasswordException;
@@ -125,8 +125,7 @@ public final class PDFExtractor extends Transform<StructuredRecord, StructuredRe
                    e.getClass().getCanonicalName(),
                    e);
         }
-      }
-      finally {
+      } finally {
         if (inputDoc != null) {
           inputDoc.close();
         }
@@ -172,7 +171,7 @@ public final class PDFExtractor extends Transform<StructuredRecord, StructuredRe
       if (inputSchema.getField(sourceFieldName).getSchema().isNullable()) {
         fieldType = inputSchema.getField(sourceFieldName).getSchema().getNonNullable().getType();
       }
-      if(fieldType != Schema.Type.BYTES) {
+      if (fieldType != Schema.Type.BYTES) {
         throw new IllegalArgumentException("Input field must be bytes but was: " +
                                              inputSchema.getField(sourceFieldName).getSchema().getType());
       }
