+++ src/test/java/io/cdap/plugin/pdf/PDFExtractorTest.java
@@ -14,13 +14,13 @@
  * the License.
  */
 
-package co.cask.hydrator.plugin;
+package io.cdap.plugin.pdf;
 
-import co.cask.cdap.api.data.format.StructuredRecord;
-import co.cask.cdap.api.data.schema.Schema;
-import co.cask.cdap.etl.api.Transform;
-import co.cask.cdap.etl.mock.common.MockEmitter;
-import co.cask.cdap.etl.mock.common.MockPipelineConfigurer;
+import io.cdap.cdap.api.data.format.StructuredRecord;
+import io.cdap.cdap.api.data.schema.Schema;
+import io.cdap.cdap.etl.api.Transform;
+import io.cdap.cdap.etl.mock.common.MockEmitter;
+import io.cdap.cdap.etl.mock.common.MockPipelineConfigurer;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
